pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "gunicorn>=23.0.0",
    "oauthlib>=3.3.1",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
]
